requirements.txt
Flask==2.0.2
werkzeug==2.0.2
pandas==2.2.3
gunicorn==20.1.0
scikit-learn==1.3.2
python-dotenv